+++ src/pic/particles/particle_pusher.hpp
@@ -4,11 +4,12 @@
 #include "wrapper.h"
 
 #include "field_macros.h"
+#include "particle_macros.h"
+#include "pic.h"
+
 #include "io/output.h"
 #include "meshblock/meshblock.h"
-#include "particle_macros.h"
 #include "meshblock/particles.h"
-#include "pic.h"
 #include "utils/qmath.h"
 
 namespace ntt {
@@ -42,7 +43,7 @@ namespace ntt {
         m_particles(particles),
         m_coeff(coeff),
         m_dt(dt),
-        m_ni2 { mblock.Ni2() } {}
+        m_ni2 { (int)mblock.Ni2() } {}
 
     /**
      * @brief Pusher for the forward Boris algorithm.
